+++ ui/app/src/components/NewExperimentNext/form/Kernel.tsx
@@ -14,18 +14,16 @@
  * limitations under the License.
  *
  */
+import Paper from '@/mui-extends/Paper'
+import Space from '@/mui-extends/Space'
+import { useStoreSelector } from '@/store'
 import AddCircleIcon from '@mui/icons-material/AddCircle'
 import RemoveCircleIcon from '@mui/icons-material/RemoveCircle'
 import { Box, IconButton, InputAdornment, MenuItem, Typography } from '@mui/material'
 import { Form, Formik } from 'formik'
 import { useEffect, useState } from 'react'
 
-import Paper from '@ui/mui-extends/esm/Paper'
-import Space from '@ui/mui-extends/esm/Space'
-
-import { useStoreSelector } from 'store'
-
-import { LabelField, SelectField, Submit, TextField } from 'components/FormField'
+import { LabelField, SelectField, Submit, TextField } from '@/components/FormField'
 
 import typesData from '../data/types'
 
@@ -33,7 +31,7 @@ interface KernelProps {
   onSubmit: (values: Record<string, any>) => void
 }
 
-const Kernel: React.FC<KernelProps> = ({ onSubmit }) => {
+const Kernel: ReactFCWithChildren<KernelProps> = ({ onSubmit }) => {
   const { spec } = useStoreSelector((state) => state.experiments)
 
   const initialValues = typesData.KernelChaos.spec!
@@ -64,13 +62,13 @@ const Kernel: React.FC<KernelProps> = ({ onSubmit }) => {
                 parameters: '',
                 predicate: '',
               },
-            ])
+            ]),
           )
 
         const removeFrame = (index: number) => () => {
           setFieldValue(
             'failKernRequest.callchain',
-            callchain.filter((_: any, i: number) => index !== i)
+            callchain.filter((_: any, i: number) => index !== i),
           )
         }