+++ ui/app/src/components/NewWorkflowNext/index.tsx
@@ -14,6 +14,9 @@
  * limitations under the License.
  *
  */
+import Paper from '@/mui-extends/Paper'
+import Space from '@/mui-extends/Space'
+import { useStoreDispatch, useStoreSelector } from '@/store'
 import TabPanelUnstyled from '@mui/base/TabPanelUnstyled'
 import TabUnstyled from '@mui/base/TabUnstyled'
 import TabsListUnstyled from '@mui/base/TabsListUnstyled'
@@ -24,14 +27,9 @@ import _ from 'lodash'
 import { useEffect, useRef, useState } from 'react'
 import type { ReactFlowInstance } from 'react-flow-renderer'
 
-import Paper from '@ui/mui-extends/esm/Paper'
-import Space from '@ui/mui-extends/esm/Space'
+import { loadRecentlyUsedExperiments } from '@/slices/workflows'
 
-import { useStoreDispatch, useStoreSelector } from 'store'
-
-import { loadRecentlyUsedExperiments } from 'slices/workflows'
-
-import YAML from 'components/YAML'
+import YAML from '@/components/YAML'
 
 import FunctionalNodesElements from './Elements/FunctionalNodes'
 import KubernetesElements from './Elements/Kubernetes'
@@ -79,7 +77,7 @@ const Tab = styled(TabUnstyled)(
     border-top-right-radius: 4px;
     border-bottom-right-radius: 4px;
   }
-  `
+  `,
 )
 const TabPanel = styled(TabPanelUnstyled)`
   flex-grow: 1;
@@ -124,11 +122,9 @@ export default function NewWorkflow() {
         <Space sx={{ height: '100%' }}>
           <Box display="flex" justifyContent="space-between" alignItems="center">
             <Box>
-              <Badge badgeContent="Preview" color="primary">
-                <Typography variant="h5" component="h1" fontWeight="bold">
-                  New Workflow
-                </Typography>
-              </Badge>
+              <Typography variant="h5" component="h1" fontWeight="bold">
+                New Workflow
+              </Typography>
               <Typography variant="body2">Use flowchart to create a new workflow.</Typography>
             </Box>
             <Space direction="row">