+++ ui/app/src/components/Token/index.tsx
@@ -14,21 +14,19 @@
  * limitations under the License.
  *
  */
-import { applyAPIAuthentication, resetAPIAuthentication } from 'api/interceptors'
+import { applyAPIAuthentication, resetAPIAuthentication } from '@/api/interceptors'
+import Space from '@/mui-extends/Space'
+import { getExperimentsState } from '@/openapi'
+import { useStoreDispatch, useStoreSelector } from '@/store'
 import { Form, Formik, FormikHelpers } from 'formik'
-import { getExperimentsState } from 'openapi'
 import { useIntl } from 'react-intl'
 
-import Space from '@ui/mui-extends/esm/Space'
+import { setAlert, setTokenName, setTokens } from '@/slices/globalStatus'
 
-import { useStoreDispatch, useStoreSelector } from 'store'
+import { Submit, TextField } from '@/components/FormField'
+import i18n from '@/components/T'
 
-import { setAlert, setTokenName, setTokens } from 'slices/globalStatus'
-
-import { Submit, TextField } from 'components/FormField'
-import i18n from 'components/T'
-
-import { validateName } from 'lib/formikhelpers'
+import { validateName } from '@/lib/formikhelpers'
 
 function validateToken(value: string) {
   let error
@@ -49,7 +47,7 @@ interface TokenProps {
   onSubmitCallback?: (values: TokenFormValues) => void
 }
 
-const Token: React.FC<TokenProps> = ({ onSubmitCallback }) => {
+const Token: ReactFCWithChildren<TokenProps> = ({ onSubmitCallback }) => {
   const intl = useIntl()
 
   const { tokens } = useStoreSelector((state) => state.globalStatus)
@@ -66,7 +64,7 @@ const Token: React.FC<TokenProps> = ({ onSubmitCallback }) => {
         setAlert({
           type: 'warning',
           message: i18n('settings.addToken.duplicateDesc', intl),
-        })
+        }),
       )
 
       return