+++ ui/app/src/components/YAMLEditor/index.tsx
@@ -14,48 +14,51 @@
  * limitations under the License.
  *
  */
+import Space from '@/mui-extends/Space'
+import { useStoreDispatch, useStoreSelector } from '@/store'
 import CloudDownloadOutlinedIcon from '@mui/icons-material/CloudDownloadOutlined'
 import PublishIcon from '@mui/icons-material/Publish'
 import { Box, Button } from '@mui/material'
-import 'ace-builds'
-import { Ace } from 'ace-builds'
-import 'ace-builds/src-min-noconflict/mode-yaml'
-import 'ace-builds/src-min-noconflict/theme-tomorrow'
-import 'ace-builds/src-min-noconflict/theme-tomorrow_night_eighties'
-import 'ace-builds/webpack-resolver'
+import { type Editor } from 'ace-builds'
+import 'ace-builds/src-noconflict/ace'
+import 'ace-builds/src-noconflict/mode-yaml'
+import 'ace-builds/src-noconflict/theme-tomorrow'
+import 'ace-builds/src-noconflict/theme-tomorrow_night'
 import fileDownload from 'js-file-download'
 import yaml from 'js-yaml'
-import { memo } from 'react'
-import { useState } from 'react'
+import { memo, useState } from 'react'
 import AceEditor, { IAceEditorProps } from 'react-ace'
 import { useIntl } from 'react-intl'
 
-import Space from '@ui/mui-extends/esm/Space'
+import { setConfirm } from '@/slices/globalStatus'
 
-import { useStoreDispatch, useStoreSelector } from 'store'
-
-import { setConfirm } from 'slices/globalStatus'
-
-import i18n from 'components/T'
+import i18n from '@/components/T'
 
 interface YAMLEditorProps {
   name?: string
   data?: string
-  mountEditor?: (editor: Ace.Editor) => void
+  mountEditor?: (editor: Editor) => void
   onUpdate?: (data: any) => void
   download?: boolean
   aceProps?: IAceEditorProps
 }
 
-const YAMLEditor: React.FC<YAMLEditorProps> = ({ name, data, mountEditor, onUpdate, download, aceProps }) => {
+const YAMLEditor: ReactFCWithChildren<YAMLEditorProps> = ({
+  name,
+  data,
+  mountEditor,
+  onUpdate,
+  download,
+  aceProps,
+}) => {
   const intl = useIntl()
 
   const { theme } = useStoreSelector((state) => state.settings)
   const dispatch = useStoreDispatch()
 
-  const [editor, setEditor] = useState<Ace.Editor>()
+  const [editor, setEditor] = useState<Editor>()
 
-  const handleOnLoad = (editor: Ace.Editor) => {
+  const handleOnLoad = (editor: Editor) => {
     setEditor(editor)
 
     typeof mountEditor === 'function' && mountEditor(editor)
@@ -66,7 +69,7 @@ const YAMLEditor: React.FC<YAMLEditorProps> = ({ name, data, mountEditor, onUpda
       setConfirm({
         title: `${i18n('common.update', intl)} ${name}`,
         handle: handleOnUpdate,
-      })
+      }),
     )
   }
 
@@ -84,7 +87,7 @@ const YAMLEditor: React.FC<YAMLEditorProps> = ({ name, data, mountEditor, onUpda
         height="100%"
         style={{ borderBottomLeftRadius: 4, borderBottomRightRadius: 4 }}
         mode="yaml"
-        theme={theme === 'light' ? 'tomorrow' : 'tomorrow_night_eighties'}
+        theme={theme === 'light' ? 'tomorrow' : 'tomorrow_night'}
         value={data}
         {...aceProps}
       />